split up RegionFilterLink into its own file
